+++ bakery/views/list.py
@@ -6,9 +6,7 @@ for building flat files.
 import logging
 from collections.abc import Callable
 
-from django.conf import settings
 from django.views.generic import ListView
-from fs import path
 
 from .base import BuildableMixin
 
@@ -45,6 +43,6 @@ class BuildableListView(ListView, BuildableMixin):
     def build_queryset(self) -> None:
         logger.debug("Building %s", self.build_path)
         self.request = self.create_request(self.build_path)
+        target_path = self.get_output_path(self.build_path)
         self.prep_directory(self.build_path)
-        target_path = path.join(str(settings.BUILD_DIR), self.build_path)
         self.build_file(target_path, self.get_content())